+++ .github/workflows/jd_pet.yml
@@ -1,13 +1,11 @@
 # This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
 
-name: jd_joy_feedPets
+name: jd_pet
 
 on:
   schedule:
-    - cron: '*/20 */1 * * *'
-  watch:
-    types: [started]
+    - cron: '35 23,4,10 * * *'
 jobs:
   build:
 
@@ -22,9 +20,11 @@ jobs:
       - name: npm install
         run: |
           npm install
-      - name: '运行 【宠汪汪喂食脚本】'
+      - name: '运行 【东东萌宠】'
         run: |
-          node jd_joy_feedPets.js
+          node jd_pet.js
         env:
           JD_COOKIE: ${{ secrets.JD_COOKIE }}
           PUSH_KEY: ${{ secrets.PUSH_KEY }}
+          BARK_PUSH: ${{ secrets.BARK_PUSH }}
+          FruitShareCodes: ${{ secrets.PetShareCodes }}
