+++ jd_pet_double.js
@@ -1,20 +1,43 @@
+//备份之前单账号的萌宠脚本
 const name = '东东萌宠';
 const $ = new Env(name);
-let cookiesArr = [], cookie = '', jdPetShareArr = [], isBox = false, notify, newShareCodes;
-//助力好友分享码(最多5个,否则后面的助力失败),原因:京东农场每人每天只有四次助力机会
-//此此内容是IOS用户下载脚本到本地使用，填写互助码的地方，同一京东账号的好友互助码请使用@符号隔开。
-//下面给出两个账号的填写示例（iOS只支持2个京东账号）
-let shareCodes = [ // 这个列表填入你要助力的好友的shareCode
-  //账号一的好友shareCode,不同好友的shareCode中间用@符号隔开
-  'MTAxODc2NTEzNTAwMDAwMDAwMjg3MDg2MA==@MTAxODc2NTEzMzAwMDAwMDAyNzUwMDA4MQ==@MTAxODc2NTEzMjAwMDAwMDAzMDI3MTMyOQ==@MTAxODc2NTEzNDAwMDAwMDAzMDI2MDI4MQ==@MTAxODcxOTI2NTAwMDAwMDAxOTQ3MjkzMw==',
-  //账号二的好友shareCode,不同好友的shareCode中间用@符号隔开
-  'MTAxODc2NTEzMjAwMDAwMDAzMDI3MTMyOQ==@MTAxODcxOTI2NTAwMDAwMDAyNjA4ODQyMQ==',
-]
-let message = '', subTitle = '', option = {}, UserName = '';
-let jdNotify = $.getdata('jdPetNotify');//是否关闭通知，false打开，true通知
-let jdServerNotify = true;//是否每次运行脚本后，都发送server酱微信通知提醒,默认是true【true:发送，false:不发送】
+
+// =======node.js使用说明======
+//Node.js用户请在jdCookie.js处填写京东ck;
+const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
+
+//ios等软件用户直接用NobyDa的jd cookie
+const cookie = jdCookieNode.CookieJD ? jdCookieNode.CookieJD : $.getdata('CookieJD');
+//京东接口地址
 const JD_API_HOST = 'https://api.m.jd.com/client.action';
-let goodsUrl = '', taskInfoKey = [];
+let jdNotify = $.getdata('jdPetNotify');
+let shareCodes = [ // 这个列表填入你要助力的好友的shareCode, 最多可能是5个
+  'MTAxODc2NTEzMjAwMDAwMDAzMDI3MTMyOQ==',
+  'MTAxODcxOTI2NTAwMDAwMDAyNjA4ODQyMQ==',
+  'MTAxODc2NTEzMDAwMDAwMDAwNTUwNDUxMw==',
+  'MTAxODc2NTEzOTAwMDAwMDAxODQ5MDg5NQ==',
+  'MTAxODcxOTI2NTAwMDAwMDAxOTQ3MjkzMw=='
+]
+// 添加box功能
+// 【用box订阅的好处】
+// 1️⃣脚本也可以远程挂载了。助力功能只需在box里面设置助力码。
+// 2️⃣所有脚本的cookie都可以备份，方便你迁移到其他支持box的软件。
+let isBox = false //默认没有使用box
+const boxShareCodeArr = ['jd_pet1', 'jd_pet2', 'jd_pet3', 'jd_pet4', 'jd_pet5'];
+isBox = boxShareCodeArr.some((item) => {
+  const boxShareCode = $.getdata(item);
+  return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
+});
+if (isBox) {
+  shareCodes = [];
+  for (const item of boxShareCodeArr) {
+    if ($.getdata(item)) {
+      shareCodes.push($.getdata(item));
+    }
+  }
+}
+let petInfo = null, taskInfo = null, message = '', subTitle = '', goodsUrl = '', taskInfoKey = [], option = {};
+
 //按顺序执行, 尽量先执行不消耗狗粮的任务, 避免中途狗粮不够, 而任务还没做完
 let function_map = {
   signInit: signInit, //每日签到
@@ -27,137 +50,26 @@ let function_map = {
   inviteFriendsInit: inviteFriendsInit, //邀请好友, 暂未处理
   feedReachInit: feedReachInit, //喂食10次任务  最后执行投食10次任务, 提示剩余狗粮是否够投食10次完成任务, 并询问要不要继续执行
 }
-!(async () => {
-  await requireConfig();
-  if (!cookiesArr[0]) {
-    $.msg($.name, '【提示】请先获取京东账号一cookie\n直接使用NobyDa的京东签到获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
-    return;
-  }
-  for (let i = 0; i < cookiesArr.length; i++) {
-    if (cookiesArr[i]) {
-      cookie = cookiesArr[i];
-      UserName = decodeURIComponent(cookie.match(/pt_pin=(.+?);/) && cookie.match(/pt_pin=(.+?);/)[1])
-      $.index = i + 1;
-      console.log(`\n开始【京东账号${$.index}】${UserName}\n`);
-      message = '';
-      subTitle = '';
-      goodsUrl = '';
-      taskInfoKey = [];
-      option = {};
-      await shareCodesFormat();
-      await jdPet();
-    }
-  }
-})()
-    .catch((e) => {
-      $.log('', `❌ ${$.name}, 失败! 原因: ${e}!`, '')
-    })
-    .finally(() => {
-      $.done();
-    })
-async function jdPet() {
-  //查询jd宠物信息
-  const initPetTownRes = await request('initPetTown');
-  message = `【京东账号${$.index}】${UserName}\n`;
-  if (initPetTownRes.code === '0' && initPetTownRes.resultCode === '0' && initPetTownRes.message === 'success') {
-    $.petInfo = initPetTownRes.result;
-    if ($.petInfo.userStatus === 0) {
-      $.msg(name, '【提示】此账号萌宠活动未开始，请手动去京东APP开启活动\n入口：我的->游戏与互动->查看更多', '', { "open-url": "openapp.jdmoble://" });
-      return
-    }
-    goodsUrl = $.petInfo.goodsInfo && $.petInfo.goodsInfo.goodsUrl;
-    option['media-url'] = goodsUrl;
-    // console.log(`初始化萌宠信息完成: ${JSON.stringify(petInfo)}`);
-    if ($.petInfo.petStatus === 5 && $.petInfo.showHongBaoExchangePop) {
-      option['open-url'] = "openApp.jdMobile://";
-      $.msg($.name, `【提醒⏰】${$.petInfo.goodsInfo.goodsName}已可领取`, '请去京东APP或微信小程序查看', option);
-      if ($.isNode()) {
-        await notify.sendNotify(`${$.name}奖品已可领取`, `京东账号${$.index} ${UserName}\n\n${$.petInfo.goodsInfo.goodsName}已可领取`);
-      }
-      if ($.isNode()) {
-        await notify.BarkNotify(`【提醒⏰】${$.petInfo.goodsInfo.goodsName}已可领取`, `请去京东APP或微信小程序查看`);
-      }
-      return
-    }
-    console.log(`\n【您的互助码shareCode】 ${$.petInfo.shareCode}\n`);
-    await taskInit();
-    if ($.taskInit.resultCode === '9999' || !$.taskInit.result) {
-      console.log('初始化任务异常, 请稍后再试');
-      return
-    }
-    $.taskInfo = $.taskInit.result;
 
-    await petSport();//遛弯
-    await slaveHelp();//助力好友
-    await masterHelpInit();//获取助力的信息
-    await doTask();//做日常任务
-    await feedPetsAgain();//再次投食
-    await energyCollect();//收集好感度
-    await showMsg();
-    console.log('全部任务完成, 如果帮助到您可以点下🌟STAR鼓励我一下, 明天见~');
-  } else if (initPetTownRes.code === '0'){
-    if (initPetTownRes.resultCode === '2001') {
-      $.msg($.name, '【提示】京东cookie已失效,请重新登录获取', 'https://bean.m.jd.com/', {"open-url": "https://bean.m.jd.com/"});
-      if ($.index === 1) {
-        $.setdata('', 'CookieJD');//cookie失效，故清空cookie。
-      } else if ($.index === 2){
-        $.setdata('', 'CookieJD2');//cookie失效，故清空cookie。
-      }
-      if ($.isNode()) {
-        await notify.sendNotify(`${$.name}cookie已失效`, `京东账号${$.index} ${UserName}\n\n请重新登录获取cookie`);
-      }
-      if ($.isNode()) {
-        await notify.BarkNotify(`${$.name}cookie已失效`, `京东账号${$.index} ${UserName}\n请重新登录获取cookie`);
-      }
-    } else {
-      console.log(`初始化萌宠失败:  ${initPetTownRes.message}`);
-    }
+let gen = entrance();
+gen.next();
+/**
+ * 入口函数
+ */
+function* entrance() {
+  if (!cookie) {
+    $.msg(name, '【提示】请先获取cookie\n直接使用NobyDa的京东签到获取', 'https://bean.m.jd.com/', { "open-url": "https://bean.m.jd.com/" });
+    $.done();
+    return
   }
-}
-// 收取所有好感度
-async function energyCollect() {
-  console.log('开始收取任务奖励好感度');
-  let function_id = arguments.callee.name.toString();
-  const response = await request(function_id);
-  // console.log(`收取任务奖励好感度完成:${JSON.stringify(response)}`);
-  if (response.code === '0') {
-    message += `【第${response.result.medalNum + 1}块勋章完成进度】${response.result.medalPercent}%，还需收集${response.result.needCollectEnergy}好感\n`;
-    message += `【已获得勋章】${response.result.medalNum}块，还需收集${response.result.needCollectMedalNum}块即可兑换奖品“${$.petInfo.goodsInfo.goodsName}”\n`;
-  }
-}
-//再次投食
-async function feedPetsAgain() {
-  const response = await request('initPetTown');//再次初始化萌宠
-  if (response.code === '0' && response.resultCode === '0' && response.message === 'success') {
-    $.petInfo = response.result;
-    let foodAmount = $.petInfo.foodAmount; //剩余狗粮
-    if (foodAmount - 100 >= 10) {
-      for (let i = 0; i < parseInt((foodAmount - 100) / 10); i++) {
-        const feedPetRes = await request('feedPets');
-        console.log(`投食feedPetRes`);
-        if (feedPetRes.resultCode == 0 && feedPetRes.code == 0) {
-          console.log('投食成功')
-        }
-      }
-      const response2 = await request('initPetTown');
-      $.petInfo = response2.result;
-      subTitle = $.petInfo.goodsInfo.goodsName;
-      message += `【与爱宠相识】${$.petInfo.meetDays}天\n`;
-      message += `【剩余狗粮】${$.petInfo.foodAmount}g\n`;
-    } else {
-      console.log("目前剩余狗粮：【" + foodAmount + "】g,不再继续投食,保留100g用于完成第二天任务");
-      subTitle = $.petInfo.goodsInfo.goodsName;
-      message += `【与爱宠相识】${$.petInfo.meetDays}天\n`;
-      message += `【剩余狗粮】${$.petInfo.foodAmount}g\n`;
-    }
-  } else {
-    console.log(`初始化萌宠失败:  ${JSON.stringify($.petInfo)}`);
-  }
-}
+  console.log('任务开始');
+  yield initPetTown(); //初始化萌宠
+  yield taskInit(); // 初始化任务
 
-
-async function doTask() {
-  $.taskInfo['taskList'].forEach((val) => {
+  yield petSport(); // 遛弯
+  yield slaveHelp();  // 助力, 在顶部shareCodes中填写需要助力的shareCode
+  yield masterHelpInit();//获取助力信息
+  taskInfo['taskList'].forEach((val) => {
     taskInfoKey.push(val);
   })
   // 任务开始
@@ -165,28 +77,374 @@ async function doTask() {
     if (taskInfoKey.indexOf(task_name) !== -1) {
       taskInfoKey.splice(taskInfoKey.indexOf(task_name), 1);
     }
-    if ($.taskInfo[task_name] && !$.taskInfo[task_name].finished) {
+    if (taskInfo[task_name] && !taskInfo[task_name].finished) {
       console.log('任务' + task_name + '开始');
       // yield eval(task_name + '()');
-      await function_map[task_name]();
+      yield function_map[task_name]();
     } else {
       console.log('任务' + task_name + '已完成');
     }
   }
   for (let item of taskInfoKey) {
-    console.log(`新任务 【${$.taskInfo[item].title}】 功能未开发，请反馈给脚本维护者@lxk0301\n`);
-    $.msg($.name, subTitle, `新的任务 【${$.taskInfo[item].title}】 功能未开发，请反馈给脚本维护者@lxk0301\n`, {"open-url": "https://t.me/JD_fruit_pet"})
+    console.log(`新任务 【${taskInfo[item].title}】 功能未开发，请反馈给脚本维护者@lxk0301\n`);
+    $.msg($.name, subTitle, `新的任务 【${taskInfo[item].title}】 功能未开发，请反馈给脚本维护者@lxk0301\n`, {"open-url": "https://t.me/JD_fruit_pet"})
+  }
+  yield feedPetsAgain();//所有任务做完后，检测剩余狗粮是否大于110g,大于就继续投食
+  yield energyCollect();
+  option['media-url'] = goodsUrl;
+  if (!jdNotify || jdNotify === 'false') {
+    $.msg(name, subTitle, message, option);
+  }
+  // $notify(name, subTitle, message);
+  console.log('全部任务完成, 如果帮助到您可以点下🌟STAR鼓励我一下, 明天见~');
+  $.done();
+}
+
+
+// 收取所有好感度
+function energyCollect() {
+  console.log('开始收取任务奖励好感度');
+
+  let function_id = arguments.callee.name.toString();
+  request(function_id).then(response => {
+    console.log(`收取任务奖励好感度完成:${JSON.stringify(response)}`);
+    if (response.code === '0') {
+      // message += `【第${petInfo.medalNum + 2}块勋章完成进度】：${response.result.medalPercent}%，还需投食${response.result.needCollectEnergy}g狗粮\n`;
+      // message += `【已获得勋章】${petInfo.medalNum + 1}块，还需收集${petInfo.goodsInfo.exchangeMedalNum - petInfo.medalNum - 1}块即可兑换奖品“${petInfo.goodsInfo.goodsName}”\n`;
+      message += `【第${response.result.medalNum + 1}块勋章完成进度】${response.result.medalPercent}%，还需收集${response.result.needCollectEnergy}好感\n`;
+      message += `【已获得勋章】${response.result.medalNum}块，还需收集${response.result.needCollectMedalNum}块即可兑换奖品“${petInfo.goodsInfo.goodsName}”\n`;
+    }
+    gen.next();
+  })
+}
+
+// 首次投食 任务
+function firstFeedInit() {
+  console.log('首次投食任务合并到10次喂食任务中');
+  setTimeout(() => {
+    gen.next();
+  }, 2000);
+}
+
+/**
+ * 投食10次 任务
+ */
+async function feedReachInit() {
+  console.log('投食任务开始...');
+
+  // let foodAmount = petInfo.foodAmount; //剩余狗粮
+  let finishedTimes = taskInfo.feedReachInit.hadFeedAmount / 10; //已经喂养了几次
+  let needFeedTimes = 10 - finishedTimes; //还需要几次
+  // let canFeedTimes = foodAmount / 10;
+  // if (canFeedTimes < needFeedTimes) {
+  // if (confirm('当前剩余狗粮' + foodAmount + 'g, 已不足投食' + needFeedTimes + '次, 确定要继续吗?') === false) {
+  // 	console.log('你拒绝了执行喂养十次任务');
+  // 	gen.next();
+  // }
+  // }
+
+  let tryTimes = 20; //尝试次数
+  do {
+    console.log(`还需要投食${needFeedTimes}次`);
+    let response = await feedPets();
+    console.log(`本次投食结果: ${JSON.stringify(response)}`);
+    if (response.resultCode == 0 && response.code == 0) {
+      needFeedTimes--;
+    }
+    if (response.resultCode == 3003 && response.code == 0) {
+      console.log('剩余狗粮不足, 投食结束');
+      needFeedTimes = 0;
+    }
+
+    tryTimes--;
+  } while (needFeedTimes > 0 && tryTimes > 0)
+
+  console.log('投食任务结束...');
+  gen.next();
+
+}
+
+// 遛狗, 每天次数上限10次, 随机给狗粮, 每次遛狗结束需调用getSportReward领取奖励, 才能进行下一次遛狗
+async function petSport() {
+  console.log('开始遛弯');
+
+  var times = 1;
+  var code = 0;
+  var resultCode = 0;
+
+  do {
+    let response = await request(arguments.callee.name.toString())
+    console.log(`第${times}次遛狗完成: ${JSON.stringify(response)}`);
+    resultCode = response.resultCode;
+
+    if (resultCode == 0) {
+      let sportRevardResult = await getSportReward();
+      console.log(`领取遛狗奖励完成: ${JSON.stringify(sportRevardResult)}`);
+    }
+
+    times++;
+  } while (resultCode == 0 && code == 0)
+  if (times > 1) {
+    message += '【十次遛狗】已完成\n';
+  }
+  gen.next();
+
+}
+
+/**
+ * 助力好友, 暂时支持一个好友, 需要拿到shareCode
+ * shareCode为你要助力的好友的
+ * 运行脚本时你自己的shareCode会在控制台输出, 可以将其分享给他人
+ */
+async function slaveHelp() {
+  let functionId = arguments.callee.name.toString();
+  let helpPeoples = '';
+  for (let code of shareCodes) {
+    console.log(`开始助力好友: ${code}`);
+    let response = await request(functionId, {
+      shareCode: code
+    });
+    if (response.code === '0' && response.resultCode === '0') {
+      if (response.result.helpStatus === 0) {
+        console.log('已给好友: 【' + response.result.masterNickName + '】助力');
+        helpPeoples += response.result.masterNickName + '，';
+      } else if (response.result.helpStatus === 1) {
+        // 您今日已无助力机会
+        console.log(`助力好友${response.result.masterNickName}失败，您今日已无助力机会`);
+        break;
+      } else if (response.result.helpStatus === 2) {
+        //该好友已满5人助力，无需您再次助力
+        console.log(`该好友${response.result.masterNickName}已满5人助力，无需您再次助力`);
+      }
+    } else {
+      console.log(`助理好友结果: ${response.message}`);
+    }
+  }
+  if (helpPeoples && helpPeoples.length > 0) {
+    message += `【您助力的好友】${helpPeoples.substr(0, helpPeoples.length - 1)}\n`;
+  }
+
+  gen.next();
+}
+
+
+// 领取遛狗奖励
+function getSportReward() {
+  return new Promise((rs, rj) => {
+    request(arguments.callee.name.toString()).then(response => {
+      rs(response);
+    })
+  })
+}
+
+// 浏览店铺任务, 任务可能为多个? 目前只有一个
+async function browseShopsInit() {
+  console.log('开始浏览店铺任务');
+  let times = 0;
+  let resultCode = 0;
+  let code = 0;
+
+  do {
+    let response = await request("getBrowseShopsReward");
+    console.log(`第${times}次浏览店铺结果: ${JSON.stringify(response)}`);
+    code = response.code;
+    resultCode = response.resultCode;
+    times++;
+  } while (resultCode == 0 && code == 0 && times < 5)
+
+  console.log('浏览店铺任务结束');
+  gen.next();
+}
+
+// 浏览指定店铺 任务
+function browseSingleShopInit() {
+  console.log('准备浏览指定店铺');
+  const body = {"index":0,"version":1,"type":1};
+  request("getSingleShopReward", body).then(response => {
+    console.log(`response::${JSON.stringify(response)}`);
+    if (response.code === '0' && response.resultCode === '0') {
+      const body2 = {"index":0,"version":1,"type":2};
+      request("getSingleShopReward", body2).then(response2 => {
+        console.log(`response2::${JSON.stringify(response)}`);
+        if (response2.code === '0' && response2.resultCode === '0') {
+          message += `【浏览指定店铺】获取${response2.result.reward}g\n`;
+        }
+        gen.next();
+      })
+    }
+  })
+}
+// 临时新增任务--冰淇淋会场
+function browseSingleShopInit2() {
+  console.log('准备浏览指定店铺--冰淇淋会场');
+  const body = {"index":1,"version":1,"type":1};
+  const body2 = {"index":1,"version":1,"type":2}
+  request("getSingleShopReward", body).then(response => {
+    console.log(`①点击浏览指定店铺结果: ${JSON.stringify(response)}`);
+    if (response.code === '0' && response.resultCode === '0') {
+      request("getSingleShopReward", body2).then(response2 => {
+        console.log(`②浏览指定店铺结果: ${JSON.stringify(response2)}`);
+        if (response2.code === '0' && response2.resultCode === '0') {
+          message += `【冰淇淋会场】获取狗粮${response2.result.reward}g\n`;
+        }
+      })
+    }
+    gen.next();
+  })
+}
+function browseSingleShopInit3() {
+  console.log('准备完成 去参与星品解锁计划');
+  const body = {"index":2,"version":1,"type":1};
+  const body2 = {"index":2,"version":1,"type":2}
+  request("getSingleShopReward", body).then(response => {
+    console.log(`①点击浏览指定店铺结果: ${JSON.stringify(response)}`);
+    if (response.code === '0' && response.resultCode === '0') {
+      request("getSingleShopReward", body2).then(response2 => {
+        console.log(`②浏览指定店铺结果: ${JSON.stringify(response2)}`);
+        if (response2.code === '0' && response2.resultCode === '0') {
+          message += `【去参与星品解锁计划】获取狗粮${response2.result.reward}g\n`;
+        }
+      })
+    }
+    gen.next();
+  })
+}
+// 三餐签到, 每天三段签到时间
+function threeMealInit() {
+  console.log('准备三餐签到');
+  request("getThreeMealReward").then(response => {
+    console.log(`三餐签到结果: ${JSON.stringify(response)}`);
+    if (response.code === '0' && response.resultCode === '0') {
+      message += `【定时领狗粮】获得${response.result.threeMealReward}g\n`;
+    } else {
+      message += `【定时领狗粮】${response.message}\n`;
+    }
+    gen.next();
+  })
+}
+
+// 每日签到, 每天一次
+function signInit() {
+  console.log('准备每日签到');
+  request("getSignReward").then(response => {
+    console.log(`每日签到结果: ${JSON.stringify(response)}`);
+    message += `【每日签到成功】奖励${response.result.signReward}g狗粮\n`;
+    gen.next();
+  })
+
+}
+
+// 投食
+function feedPets() {
+  console.log('开始投食');
+  return new Promise((rs, rj) => {
+    request(arguments.callee.name.toString()).then(response => {
+      rs(response);
+    })
+  })
+}
+
+//查询jd宠物信息
+function initPetTown() {
+  request(arguments.callee.name.toString()).then((response) => {
+    // console.log(`初始化萌宠信息${JSON.stringify(response)}`)
+    if (response.code === '0' && response.resultCode === '0' && response.message === 'success') {
+      petInfo = response.result;
+      if (petInfo.userStatus === 0) {
+        $.msg(name, '【提示】此账号萌宠活动未开始，请手动去京东APP开启活动\n入口：我的->游戏与互动->查看更多', '', { "open-url": "openapp.jdmoble://" });
+        $.done();
+        return
+      }
+      goodsUrl = petInfo.goodsInfo && petInfo.goodsInfo.goodsUrl;
+      // console.log(`初始化萌宠信息完成: ${JSON.stringify(petInfo)}`);
+      if (petInfo.petStatus === 5 && petInfo.showHongBaoExchangePop) {
+        option['open-url'] = "openApp.jdMobile://";
+        option['media-url'] = goodsUrl;
+        $.msg($.name, `【提醒⏰】${petInfo.goodsInfo.goodsName}已可领取`, '请去京东APP或微信小程序查看', option);
+        $.done();
+        return
+      }
+      console.log(`\n【您的互助码shareCode】 ${petInfo.shareCode}\n`);
+      gen.next();
+    } else if (response.code === '0' && response.resultCode === '2001'){
+      console.log(`初始化萌宠失败:  ${response.message}`);
+      $.setdata('', 'CookieJD');//cookie失效，故清空cookie。
+      $.msg(name, '【提示】京东cookie已失效,请重新登录获取', 'https://bean.m.jd.com/', { "open-url": "https://bean.m.jd.com/" });
+      $.done();
+    }
+  })
+
+}
+//再次投食
+async function feedPetsAgain() {
+  const response = await secondInitPetTown(); //再次初始化萌宠
+  if (response.code === '0' && response.resultCode === '0' && response.message === 'success') {
+    let secondPetInfo = response.result;
+    let foodAmount = secondPetInfo.foodAmount; //剩余狗粮
+    if (foodAmount - 100 >= 10) {
+      for (let i = 0; i < parseInt((foodAmount - 100) / 10); i++) {
+        const feedPetRes = await feedPets();
+        console.log(`投食feedPetRes`);
+        if (feedPetRes.resultCode == 0 && feedPetRes.code == 0) {
+          console.log('投食成功')
+        }
+      }
+      const response2 = await secondInitPetTown();
+      subTitle = response2.result.goodsInfo.goodsName;
+      message += `【与爱宠相识】${response2.result.meetDays}天\n`;
+      message += `【剩余狗粮】${response2.result.foodAmount}g\n`;
+    } else {
+      console.log("目前剩余狗粮：【" + foodAmount + "】g,不再继续投食,保留100g用于完成第二天任务");
+      subTitle = secondPetInfo.goodsInfo.goodsName;
+      message += `【与爱宠相识】${secondPetInfo.meetDays}天\n`;
+      message += `【剩余狗粮】${secondPetInfo.foodAmount}g\n`;
+    }
+  } else {
+    console.log(`初始化萌宠失败:  ${JSON.stringify(petInfo)}`);
+  }
+  gen.next();
+}
+// 再次查询萌宠信息
+function secondInitPetTown() {
+  console.log('开始再次初始化萌宠信息');
+  return new Promise((rs, rj) => {
+    request("initPetTown").then(response => {
+      rs(response);
+    })
+  })
+}
+// 邀请新用户
+function inviteFriendsInit() {
+  console.log('邀请新用户功能未实现');
+  if (taskInfo.inviteFriendsInit.status == 1 && taskInfo.inviteFriendsInit.inviteFriendsNum > 0) {
+    // 如果有邀请过新用户,自动领取60gg奖励
+    request('getInviteFriendsReward').then((res) => {
+      try {
+        if (res.code == 0 && res.resultCode == 0) {
+          console.log(`领取邀请新用户奖励成功,获得狗粮现有狗粮${taskInfo.inviteFriendsInit.reward}g，${res.result.foodAmount}g`);
+          message += `【邀请新用户】获取${taskInfo.inviteFriendsInit.reward}g\n`;
+        }
+        gen.next();
+      } catch (e) {
+        console.log('领取邀请新用户奖励失败')
+      }
+    });
+  } else {
+    setTimeout(() => {
+      gen.next();
+    }, 2000);
   }
 }
+
 // 好友助力信息
 async function masterHelpInit() {
   let res = await request(arguments.callee.name.toString());
-  // console.log(`助力信息: ${JSON.stringify(res)}`);
+  console.log('助力信息: ' , res);
   if (res.code === '0' && res.resultCode === '0') {
     if (res.result.masterHelpPeoples && res.result.masterHelpPeoples.length >= 5) {
       if(!res.result.addedBonusFlag) {
         console.log("开始领取额外奖励");
-        let getHelpAddedBonusResult = await request('getHelpAddedBonus');
+        let getHelpAddedBonusResult = await getHelpAddedBonus();
         console.log(`领取30g额外奖励结果：【${getHelpAddedBonusResult.message}】`);
         message += `【额外奖励${getHelpAddedBonusResult.result.reward}领取】${getHelpAddedBonusResult.message}\n`;
       } else {
@@ -210,270 +468,34 @@ async function masterHelpInit() {
       message += `【助力您的好友】${str}\n`;
     }
   }
+  gen.next();
 }
-/**
- * 助力好友, 暂时支持一个好友, 需要拿到shareCode
- * shareCode为你要助力的好友的
- * 运行脚本时你自己的shareCode会在控制台输出, 可以将其分享给他人
- */
-async function slaveHelp() {
-  let helpPeoples = '';
-  for (let code of newShareCodes) {
-    console.log(`开始助力好友: ${code}`);
-    let response = await request(arguments.callee.name.toString(), {'shareCode': code});
-    if (response.code === '0' && response.resultCode === '0') {
-      if (response.result.helpStatus === 0) {
-        console.log('已给好友: 【' + response.result.masterNickName + '】助力');
-        helpPeoples += response.result.masterNickName + '，';
-      } else if (response.result.helpStatus === 1) {
-        // 您今日已无助力机会
-        console.log(`助力好友${response.result.masterNickName}失败，您今日已无助力机会`);
-        break;
-      } else if (response.result.helpStatus === 2) {
-        //该好友已满5人助力，无需您再次助力
-        console.log(`该好友${response.result.masterNickName}已满5人助力，无需您再次助力`);
-      }
-    } else {
-      console.log(`助理好友结果: ${response.message}`);
-    }
-  }
-  if (helpPeoples && helpPeoples.length > 0) {
-    message += `【您助力的好友】${helpPeoples.substr(0, helpPeoples.length - 1)}\n`;
-  }
+// 领取5好友助力后的奖励
+function getHelpAddedBonus() {
+  return new Promise((rs, rj)=> {
+    request(arguments.callee.name.toString()).then(response=> {
+      rs(response);
+    })
+  })
 }
-// 遛狗, 每天次数上限10次, 随机给狗粮, 每次遛狗结束需调用getSportReward领取奖励, 才能进行下一次遛狗
-async function petSport() {
-  console.log('开始遛弯');
-  let times = 1
-  const code = 0
-  let resultCode = 0
-  do {
-    let response = await request(arguments.callee.name.toString())
-    console.log(`第${times}次遛狗完成: ${JSON.stringify(response)}`);
-    resultCode = response.resultCode;
-    if (resultCode == 0) {
-      let sportRevardResult = await request('getSportReward');
-      console.log(`领取遛狗奖励完成: ${JSON.stringify(sportRevardResult)}`);
-    }
-    times++;
-  } while (resultCode == 0 && code == 0)
-  if (times > 1) {
-    message += '【十次遛狗】已完成\n';
-  }
-}
+
 // 初始化任务, 可查询任务完成情况
-async function taskInit() {
+function taskInit() {
   console.log('开始任务初始化');
-  $.taskInit = await request(arguments.callee.name.toString(), {"version":1});
-}
-// 每日签到, 每天一次
-async function signInit() {
-  console.log('准备每日签到');
-  const response = await request("getSignReward");
-  console.log(`每日签到结果: ${JSON.stringify(response)}`);
-  message += `【每日签到成功】奖励${response.result.signReward}g狗粮\n`;
-}
-
-// 三餐签到, 每天三段签到时间
-async function threeMealInit() {
-  console.log('准备三餐签到');
-  const response = await request("getThreeMealReward");
-  console.log(`三餐签到结果: ${JSON.stringify(response)}`);
-  if (response.code === '0' && response.resultCode === '0') {
-    message += `【定时领狗粮】获得${response.result.threeMealReward}g\n`;
-  } else {
-    message += `【定时领狗粮】${response.message}\n`;
-  }
-}
-
-// 浏览指定店铺 任务
-async function browseSingleShopInit() {
-  console.log('准备浏览指定店铺');
-  const body = {"index":0,"version":1,"type":1};
-  const response = await request("getSingleShopReward", body);
-  console.log(`点击进去response::${JSON.stringify(response)}`);
-  if (response.code === '0' && response.resultCode === '0') {
-    const body2 = {"index":0,"version":1,"type":2};
-    const response2 = await request("getSingleShopReward", body2);
-    console.log(`浏览完毕领取奖励:response2::${JSON.stringify(response2)}`);
-    if (response2.code === '0' && response2.resultCode === '0') {
-      message += `【浏览指定店铺】获取${response2.result.reward}g\n`;
+  const body = {"version":1};
+  request(arguments.callee.name.toString(), body).then(response => {
+    if (response.resultCode === '9999' || !response.result) {
+      console.log('初始化任务异常, 请稍后再试');
+      gen.return();
     }
-  }
-}
-// 临时新增任务--冰淇淋会场
-async function browseSingleShopInit2() {
-  console.log('准备浏览指定店铺--冰淇淋会场');
-  const body = {"index":1,"version":1,"type":1};
-  const body2 = {"index":1,"version":1,"type":2}
-  const response = await request("getSingleShopReward", body);
-  if (response.code === '0' && response.resultCode === '0') {
-    const response2 = await request("getSingleShopReward", body2);
-    console.log(`②浏览指定店铺结果: ${JSON.stringify(response2)}`);
-    if (response2.code === '0' && response2.resultCode === '0') {
-      message += `【冰淇淋会场】获取狗粮${response2.result.reward}g\n`;
-    }
-  }
-}
-async function browseSingleShopInit3() {
-  console.log('准备完成 去参与星品解锁计划');
-  const body = {"index":2,"version":1,"type":1};
-  const body2 = {"index":2,"version":1,"type":2};
-  const response = await request("getSingleShopReward", body);
-  if (response.code === '0' && response.resultCode === '0') {
-    const response2 = await request("getSingleShopReward", body2);
-    console.log(`②浏览指定店铺结果: ${JSON.stringify(response2)}`);
-    if (response2.code === '0' && response2.resultCode === '0') {
-      message += `【去参与星品解锁计划】获取狗粮${response2.result.reward}g\n`;
-    }
-  }
-}
-
-// 浏览店铺任务, 任务可能为多个? 目前只有一个
-async function browseShopsInit() {
-  console.log('开始浏览店铺任务');
-  let times = 0;
-  let resultCode = 0;
-  let code = 0;
-  do {
-    let response = await request("getBrowseShopsReward");
-    console.log(`第${times}次浏览店铺结果: ${JSON.stringify(response)}`);
-    code = response.code;
-    resultCode = response.resultCode;
-    times++;
-  } while (resultCode == 0 && code == 0 && times < 5)
-  console.log('浏览店铺任务结束');
-}
-// 首次投食 任务
-function firstFeedInit() {
-  console.log('首次投食任务合并到10次喂食任务中\n');
-}
-
-// 邀请新用户
-async function inviteFriendsInit() {
-  console.log('邀请新用户功能未实现');
-  if ($.taskInfo.inviteFriendsInit.status == 1 && $.taskInfo.inviteFriendsInit.inviteFriendsNum > 0) {
-    // 如果有邀请过新用户,自动领取60gg奖励
-    const res = await request('getInviteFriendsReward');
-    if (res.code == 0 && res.resultCode == 0) {
-      console.log(`领取邀请新用户奖励成功,获得狗粮现有狗粮${$.taskInfo.inviteFriendsInit.reward}g，${res.result.foodAmount}g`);
-      message += `【邀请新用户】获取狗粮${$.taskInfo.inviteFriendsInit.reward}g\n`;
-    }
-  }
-}
-
-/**
- * 投食10次 任务
- */
-async function feedReachInit() {
-  console.log('投食任务开始...');
-  let finishedTimes = $.taskInfo.feedReachInit.hadFeedAmount / 10; //已经喂养了几次
-  let needFeedTimes = 10 - finishedTimes; //还需要几次
-  let tryTimes = 20; //尝试次数
-  do {
-    console.log(`还需要投食${needFeedTimes}次`);
-    const response = await request('feedPets');
-    console.log(`本次投食结果: ${JSON.stringify(response)}`);
-    if (response.resultCode == 0 && response.code == 0) {
-      needFeedTimes--;
-    }
-    if (response.resultCode == 3003 && response.code == 0) {
-      console.log('剩余狗粮不足, 投食结束');
-      needFeedTimes = 0;
-    }
-    tryTimes--;
-  } while (needFeedTimes > 0 && tryTimes > 0)
-  console.log('投食任务结束...\n');
-}
-async function showMsg() {
-  if (!jdNotify || jdNotify === 'false') {
-    $.msg($.name, subTitle, message, option);
-    const notifyMessage = message.replace(/[\n\r]/g, '\n\n');
-    if (jdServerNotify) {
-      if ($.isNode()) {
-        await notify.sendNotify(`${$.name} - 账号${$.index} - ${UserName}`, `${subTitle}\n\n${notifyMessage}`);
-      }
-      if ($.isNode()) {
-        await notify.BarkNotify(`${$.name}`, `${subTitle}\n${message}`);
-      }
-    }
-  }
-}
-function shareCodesFormat() {
-  return new Promise(resolve => {
-    console.log(`第${$.index}个京东账号的助力码:::${jdPetShareArr[$.index - 1]}`)
-    if (jdPetShareArr[$.index - 1]) {
-      newShareCodes = jdPetShareArr[$.index - 1].split('@');
-    } else {
-      console.log(`由于您未提供shareCode,将采纳本脚本自带的助力码\n`)
-      const tempIndex = $.index > shareCodes.length ? (shareCodes.length - 1) : ($.index - 1);
-      newShareCodes = shareCodes[tempIndex].split('@');
-    }
-    console.log(`格式化后第${$.index}个京东账号的助力码${JSON.stringify(newShareCodes)}`)
-    resolve();
+    taskInfo = response.result;
+    // function_map = taskInfo.taskList;
+    // console.log(`任务初始化完成: ${JSON.stringify(taskInfo)}`);
+    gen.next();
   })
+
 }
-function requireConfig() {
-  return new Promise(resolve => {
-    console.log('开始获取东东萌宠配置文件\n')
-    notify = $.isNode() ? require('./sendNotify') : '';
-    //Node.js用户请在jdCookie.js处填写京东ck;
-    const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
-    const jdPetShareCodes = $.isNode() ? require('./jdPetShareCodes.js') : '';
-    //IOS等用户直接用NobyDa的jd cookie
-    if ($.isNode()) {
-      Object.keys(jdCookieNode).forEach((item) => {
-        if (jdCookieNode[item]) {
-          cookiesArr.push(jdCookieNode[item])
-        }
-      })
-    } else {
-      cookiesArr.push($.getdata('CookieJD'));
-      cookiesArr.push($.getdata('CookieJD2'));
-    }
-    console.log(`共${cookiesArr.length}个京东账号\n`)
-    if ($.isNode()) {
-      Object.keys(jdPetShareCodes).forEach((item) => {
-        if (jdPetShareCodes[item]) {
-          jdPetShareArr.push(jdPetShareCodes[item])
-        }
-      })
-    } else {
-      const boxShareCodeArr = ['jd_pet1', 'jd_pet2', 'jd_pet3', 'jd_pet4', 'jd_pet5'];
-      const boxShareCodeArr2 = ['jd2_pet1', 'jd2_pet2', 'jd2_pet3', 'jd2_pet4', 'jd2_pet5'];
-      const isBox1 = boxShareCodeArr.some((item) => {
-        const boxShareCode = $.getdata(item);
-        return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
-      });
-      const isBox2 = boxShareCodeArr2.some((item) => {
-        const boxShareCode = $.getdata(item);
-        return (boxShareCode !== undefined && boxShareCode !== null && boxShareCode !== '');
-      });
-      isBox = isBox1 ? isBox1 : isBox2;
-      if (isBox1) {
-        let temp = [];
-        for (const item of boxShareCodeArr) {
-          if ($.getdata(item)) {
-            temp.push($.getdata(item))
-          }
-        }
-        jdPetShareArr.push(temp.join('@'));
-      }
-      if (isBox2) {
-        let temp = [];
-        for (const item of boxShareCodeArr2) {
-          if ($.getdata(item)) {
-            temp.push($.getdata(item))
-          }
-        }
-        jdPetShareArr.push(temp.join('@'));
-      }
-    }
-    console.log(`jdPetShareArr::${JSON.stringify(jdPetShareArr)}`)
-    console.log(`jdPetShareArr账号长度::${jdPetShareArr.length}`)
-    resolve()
-  })
-}
+
 // 请求
 async function request(function_id, body = {}) {
   await $.wait(3000); //歇口气儿, 不然会报操作频繁
@@ -481,19 +503,19 @@ async function request(function_id, body = {}) {
     $.get(taskurl(function_id, body), (err, resp, data) => {
       try {
         if (err) {
-          console.log('\n东东萌宠: API查询请求失败 ‼️‼️');
-          $.logErr(err);
+          console.log('\n东东萌宠: API查询请求失败 ‼️‼️')
         } else {
           data = JSON.parse(data);
         }
       } catch (e) {
-        $.logErr(e, resp);
+        console.log(e)
       } finally {
         resolve(data)
       }
     })
   })
 }
+
 function taskurl(function_id, body = {}) {
   return {
     url: `${JD_API_HOST}?functionId=${function_id}&appid=wh5&loginWQBiz=pet-town&body=${escape(JSON.stringify(body))}`,
